requirements.txt
transformers
PyPDF2
python-docx
flask
boto3
torch
